fix: Make Stats page admin-only as per requirements
User feedback: "Stats must be only on the administration part, put in the admin sidebar only"

Changes:
- Moved stats page from /stats to /admin/stats
- Removed Stats from regular user navigation (Navigation.tsx getNavigationItems)
- Removed Stats from TopNavbar (regular user navigation)
- Added Stats to manager navigation (Navigation.tsx getManagerItems)
- Added Stats to admin navigation (Navigation.tsx getAdminItems)
- Simplified Stats page to only show admin statistics
- Removed user-specific stats logic since only admins access this page
