+++ frontend/src/app/(dashboard)/admin/stats/page.tsx
@@ -1,0 +1,206 @@
+'use client';
+
+import React, { useEffect, useState } from 'react';
+import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
+import { dashboardApi } from '@/lib/api';
+import { useTranslation } from '@/lib/i18n';
+
+interface DashboardStats {
+  totalUsers: number;
+  pendingUsers: number;
+  activeUsers: number;
+  totalDepartments: number;
+  activeDepartments: number;
+  totalJobPositions: number;
+  recentActivity: ActivityItem[];
+  usersByRole: Record<string, number>;
+  jobPositionsByLevel: Record<string, number>;
+  departmentHierarchy: {
+    rootDepartments: number;
+    withSubDepartments: number;
+  };
+}
+
+interface ActivityItem {
+  id: string;
+  type: 'user_registered' | 'user_approved' | 'user_rejected' | 'department_created';
+  description: string;
+  timestamp: string;
+  actor?: string;
+}
+
+export default function StatsPage() {
+  const { t } = useTranslation('dashboard');
+  const [stats, setStats] = useState<DashboardStats>({
+    totalUsers: 0,
+    pendingUsers: 0,
+    activeUsers: 0,
+    totalDepartments: 0,
+    activeDepartments: 0,
+    totalJobPositions: 0,
+    recentActivity: [],
+    usersByRole: {},
+    jobPositionsByLevel: {},
+    departmentHierarchy: {
+      rootDepartments: 0,
+      withSubDepartments: 0,
+    },
+  });
+  const [loading, setLoading] = useState(true);
+
+  useEffect(() => {
+    const fetchDashboardData = async () => {
+      try {
+        const dashboardStats = await dashboardApi.getStats();
+        setStats(dashboardStats);
+      } catch (error) {
+        console.error('Failed to fetch dashboard data:', error);
+        setStats({
+          totalUsers: 0,
+          pendingUsers: 0,
+          activeUsers: 0,
+          totalDepartments: 0,
+          activeDepartments: 0,
+          totalJobPositions: 0,
+          recentActivity: [],
+          usersByRole: {},
+          jobPositionsByLevel: {},
+          departmentHierarchy: {
+            rootDepartments: 0,
+            withSubDepartments: 0,
+          },
+        });
+      } finally {
+        setLoading(false);
+      }
+    };
+
+    fetchDashboardData();
+  }, []);
+
+  if (loading) {
+    return (
+      <div className="p-8">
+        <div className="animate-pulse">
+          <div className="h-8 bg-gray-200 rounded w-1/4 mb-6"></div>
+          <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6 mb-8">
+            {[...Array(6)].map((_, i) => (
+              <div key={i} className="h-32 bg-gray-200 rounded-lg"></div>
+            ))}
+          </div>
+        </div>
+      </div>
+    );
+  }
+
+  return (
+    <div className="p-8">
+      <div className="mb-8">
+        <h1 className="text-3xl font-bold text-gray-900">
+          {t('stats.title', { defaultValue: 'Statistics' })}
+        </h1>
+        <p className="text-gray-600 mt-2">
+          {t('stats.subtitle', { defaultValue: 'System-wide statistics and metrics' })}
+        </p>
+      </div>
+
+      {/* Admin Stats Grid */}
+      <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6">
+        {/* Total Users */}
+        <Card>
+          <CardHeader className="flex flex-row items-center justify-between space-y-0 pb-2">
+            <CardTitle className="text-sm font-medium">{t('stats.totalUsers')}</CardTitle>
+            <svg className="w-4 h-4 text-blue-600" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+              <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M12 4.354a4 4 0 110 5.292M15 21H3v-1a6 6 0 0112 0v1zm0 0h6v-1a6 6 0 00-9-5.197m13.5-1a4 4 0 11-8 0 4 4 0 018 0z" />
+            </svg>
+          </CardHeader>
+          <CardContent>
+            <div className="text-2xl font-bold">{stats.totalUsers}</div>
+            <p className="text-xs text-gray-600">
+              <span className="text-green-600">{t('stats.activeCount', { count: stats.activeUsers })}</span> •
+              <span className="text-yellow-600 ml-1">{t('stats.pendingCount', { count: stats.pendingUsers })}</span>
+            </p>
+          </CardContent>
+        </Card>
+
+        {/* Pending Approvals */}
+        <Card>
+          <CardHeader className="flex flex-row items-center justify-between space-y-0 pb-2">
+            <CardTitle className="text-sm font-medium">{t('stats.pendingApprovals')}</CardTitle>
+            <svg className="w-4 h-4 text-yellow-600" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+              <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M12 8v4l3 3m6-3a9 9 0 11-18 0 9 9 0 0118 0z" />
+            </svg>
+          </CardHeader>
+          <CardContent>
+            <div className="text-2xl font-bold">{stats.pendingUsers}</div>
+            <p className="text-xs text-gray-600">
+              {t('stats.pendingApprovalsDesc')}
+            </p>
+          </CardContent>
+        </Card>
+
+        {/* Active Users */}
+        <Card>
+          <CardHeader className="flex flex-row items-center justify-between space-y-0 pb-2">
+            <CardTitle className="text-sm font-medium">{t('stats.activeUsers')}</CardTitle>
+            <svg className="w-4 h-4 text-green-600" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+              <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M9 12l2 2 4-4m6 2a9 9 0 11-18 0 9 9 0 0118 0z" />
+            </svg>
+          </CardHeader>
+          <CardContent>
+            <div className="text-2xl font-bold">{stats.activeUsers}</div>
+            <p className="text-xs text-gray-600">
+              {t('stats.activeUsersDesc')}
+            </p>
+          </CardContent>
+        </Card>
+
+        {/* Departments */}
+        <Card>
+          <CardHeader className="flex flex-row items-center justify-between space-y-0 pb-2">
+            <CardTitle className="text-sm font-medium">{t('stats.departments')}</CardTitle>
+            <svg className="w-4 h-4 text-purple-600" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+              <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M19 21V5a2 2 0 00-2-2H7a2 2 0 00-2 2v16m14 0h2m-2 0h-5m-9 0H3m2 0h5M9 7h1m-1 4h1m4-4h1m-1 4h1m-5 10v-5a1 1 0 011-1h2a1 1 0 011 1v5m-4 0h4" />
+            </svg>
+          </CardHeader>
+          <CardContent>
+            <div className="text-2xl font-bold">{stats.totalDepartments}</div>
+            <p className="text-xs text-gray-600">
+              <span className="text-green-600">{t('stats.activeCount', { count: stats.activeDepartments })}</span>
+            </p>
+          </CardContent>
+        </Card>
+
+        {/* Job Positions */}
+        <Card>
+          <CardHeader className="flex flex-row items-center justify-between space-y-0 pb-2">
+            <CardTitle className="text-sm font-medium">{t('stats.jobPositions')}</CardTitle>
+            <svg className="w-4 h-4 text-indigo-600" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+              <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M21 13.255A23.931 23.931 0 0112 15c-3.183 0-6.22-.62-9-1.745M16 6V4a2 2 0 00-2-2h-4a2 2 0 00-2-2v2m8 0V6a2 2 0 012 2v6a2 2 0 01-2 2H8a2 2 0 01-2-2V8a2 2 0 012-2V6" />
+            </svg>
+          </CardHeader>
+          <CardContent>
+            <div className="text-2xl font-bold">{stats.totalJobPositions}</div>
+            <p className="text-xs text-gray-600">
+              {t('stats.jobPositionsDesc')}
+            </p>
+          </CardContent>
+        </Card>
+
+        {/* System Status */}
+        <Card>
+          <CardHeader className="flex flex-row items-center justify-between space-y-0 pb-2">
+            <CardTitle className="text-sm font-medium">{t('stats.systemStatus')}</CardTitle>
+            <div className="w-2 h-2 bg-green-500 rounded-full"></div>
+          </CardHeader>
+          <CardContent>
+            <div className="text-2xl font-bold text-green-600">{t('stats.healthy')}</div>
+            <p className="text-xs text-gray-600">
+              {t('stats.allSystemsOperational')}
+            </p>
+          </CardContent>
+        </Card>
+      </div>
+    </div>
+  );
+}
