+++ frontend/src/components/layout/TopNavbar.tsx
@@ -29,15 +29,6 @@ const getUserNavigationItems = (t: any) => [
     ),
   },
   {
-    name: t('navigation.stats'),
-    href: '/stats',
-    icon: (
-      <svg className="w-5 h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-        <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M9 19v-6a2 2 0 00-2-2H5a2 2 0 00-2 2v6a2 2 0 002 2h2a2 2 0 002-2zm0 0V9a2 2 0 012-2h2a2 2 0 012 2v10m-6 0a2 2 0 002 2h2a2 2 0 00 2-2m0 0V5a2 2 0 012-2h2a2 2 0 012 2v14a2 2 0 01-2 2h-2a2 2 0 01-2-2z" />
-      </svg>
-    ),
-  },
-  {
     name: t('navigation.documents') || 'Documents',
     href: '/documents',
     icon: (
